requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-cloud-vision
openai
pandas
scikit-learn
gspread
oauth2client
pillow
pdf2image
python-telegram-bot
httpx
uuid
